refactor: expose custom triggers and executor interface discord
