+++ packages/plugin-discord/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./plugin";
 export * from "./types";
 export * from "./templates";
+export * from "./executors";
+export * from "./triggers";
